Calculate Confounder-Adjusted Median Survival Times

This function can be utilized to calculate the confounder-adjusted median survival time, given previously estimated adjusted survival curves.

Usage

adjusted_median_survival(adjsurv, p=0.5, verbose=TRUE,
                         interpolation="steps")

Arguments

adjsurv

An adjustedsurv object created using the adjustedsurv function.

p

The quantile of interest. To calculate the median survival time, set this parameter to 0.5 (default).

verbose

Whether to print the output or not.

interpolation

Either "steps" (default) or "linear". This parameter controls how interpolation is performed. If this argument is set to "steps", the curves will be treated as step functions. If it is set to "linear", the curves wil be treated as if there are straight lines between the point estimates instead. Points that lie between estimated points will be interpolated accordingly.

Details

The median survival time is simply the time when half the patients are expected to be alive. The chance of surviving beyond that time is 50 percent. It can be read directly from the respective survival curve by drawing a straight line from the 0.5 point on the Y-Axis and reading the X-Axis value where this line intersects with the survival curve. If the survival probability never drops below 50 percent, the median survival time cannot be calculated. By changing the p argument to some other value between 0 and 1 the user can calculate the same quantity for different quantiles.

This function calculates this quantity automatically. When multiple imputation was used in the original function call, the median survival times are read off the final pooled survival curves directly.

Value

Silently returns the data.frame object which can be seen when verbose=TRUE. The column median_surv includes the confounder-adjusted median survival times.

Author

Robin Denz

See also

Examples

if (FALSE) {
library(survival)

# simulate some data as example
sim_dat <- sim_confounded_surv(n=500, max_t=1.2)
sim_dat$group <- as.factor(sim_dat$group)

# estimate a cox-regression for the outcome
cox_mod <- coxph(Surv(time, event) ~ x1 + x2 + x3 + x4 + x5 + x6 + group,
                 data=sim_dat, x=TRUE)

# use it to calculate adjusted survival curves with bootstrapping
adjsurv <- adjustedsurv(data=sim_dat,
                        variable="group",
                        ev_time="time",
                        event="event",
                        method="direct",
                        outcome_model=cox_mod,
                        conf_int=FALSE,
                        bootstrap=FALSE)

# calculate adjusted median survival times
adjusted_median_survival(adjsurv)

}