Plot the Difference of Two Adjusted Survival Curves or CIFs

A function to graphically display the difference between two confounder-adjusted survival curves which where previously estimated using the adjustedsurv function or between two confounder-adjusted CIFs which where previously estimated using the adjustedcif function. The user can customize the plot using a variety of options. Internally it uses the ggplot2 package, so additional not implemented features can be added using the standard ggplot2 syntax.

Usage

plot_difference(x, group_1=NULL, group_2=NULL,
                conf_int=FALSE, conf_level=0.95,
                type="steps", max_t=Inf,
                size=0.7, color="black", linetype="solid",
                alpha=1, conf_int_alpha=0.4,
                points_ci_size=NULL, points_ci_width=NULL,
                xlab="Time", ylab=NULL, title=NULL,
                subtitle=NULL, gg_theme=ggplot2::theme_classic(),
                line_at_0=TRUE, line_at_0_size=0.7,
                line_at_0_color="grey", line_at_0_linetype="dashed",
                line_at_0_alpha=1,
                loess_smoother=FALSE, loess_span=0.75,
                loess_color=color, loess_size=size,
                loess_linetype="dashed", loess_alpha=alpha,
                ...)

Arguments

x

An adjustedsurv object created using the adjustedsurv function or an adjustedcif object created using the adjustedcif function.

group_1

A single character string specifying one of the levels of the variable used in the original adjustedsurv or adjustedcif function call. This group will be subtracted from. For example if group_1="A" and group_2="B" the plotted curve will correspond to the survival probability of A minus the survival probability of B over time.

group_2

Also a single character string specifying one of the levels of variable. This corresponds to the right side of the difference equation. See argument group_2.

conf_int

A logical variable indicating whether the confidence intervals should be drawn. This only works when bootstrap=TRUE was used in the original adjustedsurv or adjustedcif function call. It also currently does not work if multiply imputed data was used.

conf_level

The confidence level that should be used when calculating the confidence intervals. Ignored if conf_int=FALSE.

type

Must be one of "steps" (drawing the difference as a step function), "lines" (drawing the difference using linear interpolation), "points" (drawing points only) or "none" (drawing nothing, useful when only the smoothed difference is of interest). It defaults to "steps".

max_t

A number indicating the latest time to which the curve should be extended to.

size

A number controlling the thickness of the difference curve.

color

A string specifying the color of the difference curve.

linetype

A string specifying the linetype of the difference curve.

alpha

A number controlling the transparency level of the difference curves.

conf_int_alpha

A number indicating the level of transparency that should be used when drawing the confidence regions.

points_ci_size

Only used when type="points". Controls the size of the error bars.

points_ci_width

Only used when type="points". Controls the width of the error bars.

xlab

A character string to be used as the X-Axis label of the plot. Defaults to "Time".

ylab

A character string to be used as the Y-Axis label of the plot. By default (NULL) uses the equation used to calculate the differences, based on the names supplied in group_1 and group_2.

title

A character string to be used as the title of the plot. Set to NULL if no title should be used.

subtitle

A character string to be used as the subtitle of the plot. Set to NULL if no subtitle should be used.

gg_theme

A ggplot2 theme object which will be used for the plot.

line_at_0

Whether to draw a horizontal line at y = 0 or not.

line_at_0_size

The size of the line drawn at y = 0. Ignored if line_at_0=FALSE.

line_at_0_color

The color of the line drawn at y = 0. Ignored if line_at_0=FALSE.

line_at_0_linetype

The linetype of the line drawn at y = 0. Ignored if line_at_0=FALSE.

line_at_0_alpha

The transparency level of the line drawn at y = 0. Ignored if line_at_0=FALSE.

loess_smoother

Whether to draw a LOESS smoother through the difference curves.

loess_span

The span of the LOESS smoother. Ignored if loess_smoother=FALSE. See stat_smooth in the ggplot2 package, method="loess" for more details.

loess_color

The color of the LOESS smoother line. Ignored if loess_smoother=FALSE.

loess_size

The size of the LOESS smoother line. Ignored if loess_smoother=FALSE.

loess_linetype

The linetype of the LOESS smoother line. Ignored if loess_smoother=FALSE.

loess_alpha

The transparency level of the LOESS smoother line. Ignored if loess_smoother=FALSE.

...

Currently not used.

Details

This function allows the easy creation of difference curves. The syntax is exactly the same for both adjusted survival curves and adjusted CIFs. It currently does not support plotting multiple difference curves at once, which could be useful when there are more than two treatment groups in variable. If the user is interested in this, we recommend calling this function multiple times with the desired comparisons and concatenating the individual plots into one plot afterwards using a suitable function such as par or ggarrange.

Value

Returns a ggplot2 object.

Author

Robin Denz

See also

Examples

if (FALSE) {
library(survival)
library(ggplot2)

# simulate some data as example
sim_dat <- sim_confounded_surv(n=500, max_t=1.2)
sim_dat$group <- as.factor(sim_dat$group)

# estimate a cox-regression for the outcome
cox_mod <- coxph(Surv(time, event) ~ x1 + x2 + x3 + x4 + x5 + x6 + group,
                 data=sim_dat, x=TRUE)

# use it to calculate adjusted survival curves with bootstrapping
adjsurv <- adjustedsurv(data=sim_dat,
                        variable="group",
                        ev_time="time",
                        event="event",
                        method="direct",
                        outcome_model=cox_mod,
                        conf_int=TRUE,
                        bootstrap=TRUE,
                        n_boot=500)

# plot the difference with default values
plot_difference(adjsurv)

# plot with reversed differences
plot_difference(adjsurv, group_1="1", group_2="0")

# plot with confidence intervals
plot_difference(adjsurv, conf_int=TRUE)

# plot using lines instead
plot_difference(adjsurv, conf_int=TRUE, type="lines")

# plot using points instead
plot_difference(adjsurv, conf_int=TRUE, type="points")

# plot using an additional loess smoother
plot_difference(adjsurv, loess_smoother=TRUE)

# plot without the line at 0
plot_difference(adjsurv, line_at_0=FALSE)

# plot with some custom parameters
plot_difference(adjsurv, conf_int=TRUE, color="blue", linetype="dotted",
                alpha=0.8, line_at_0_size=1.1, line_at_0_color="red",
                loess_smoother=TRUE, loess_span=0.55)

}