Targeted Maximum Likelihood Estimation for Adjusted Survival Curves

This page explains the details of estimating adjusted survival curves using the targeted maximum likelihood methodology for single event time-to-event data (method="tmle" in the adjustedsurv function). All regular arguments of the adjustedsurv function can be used. Additionally, you have to specify either SL.ftime or glm.ftime, SL.ctime or glm.ctime and SL.trt or glm.trt in the adjustedsurv call. Further arguments specific to this method are listed below.

Arguments

adjust_vars

A character vector of column names specifying variables to be used when modeling the outcome, treatment and censoring mechanism. Can be set to NULL (default), in which case all covariates are used. See details and examples.

SL.ftime

A character vector or list specification to be passed to the SL.library option in the call to SuperLearner for the outcome regression. See?SuperLearner for more information on how to specify valid SuperLearner libraries. It is expected that the wrappers used in the library will play nicely with the input variables, which will be called "trt", names(adjust_vars), and "t".

SL.ctime

A character vector or list specification to be passed to the SL.library argument in the call to SuperLearner for the estimate of the conditional hazard for censoring. It is expected that the wrappers used in the library will play nicely with the input variables, which will be called "trt" and names(adjust_vars).

SL.trt

A character vector or list specification to be passed to the SL.library argument in the call to SuperLearner for the estimate of the conditional probability of treatment. It is expected that the wrappers used in the library will play nicely with the input variables, which will be names(adjust_vars).

glm.ftime

A character specification of the right-hand side of the equation passed to the formula option of a call to glm for the outcome regression. Ignored if SL.ftime is not equal to NULL. Use "trt" to specify the treatment in this formula (see examples). The formula can additionally include any variables found in names(adjust_vars).

glm.ctime

A character specification of the right-hand side of the equation passed to the formula option of a call to glm for the estimate of the conditional hazard for censoring. Ignored if SL.ctime is not equal to NULL. Use "trt" to specify the treatment in this formula (see examples). The formula can additionally include any variables found in names(adjust_vars).

glm.trt

A character specification of the right-hand side of the equation passed to the formula option of a call to glm for the estimate of the conditional probability of treatment. Ignored if SL.trt is not equal to NULL. The formula can include any variables found in names(adjust_vars).

...

Additional arguments passed to survtmle.

Details

Type of Adjustment: Adjustments are made based on the treatment assignment mechanism, the outcome mechanism and the censoring mechanism. No models can be supplied. The adjustments are made based on SuperLearner libraries or using the glm arguments.

Doubly-Robust: Estimates are Doubly-Robust.

Categorical groups: Currently only two groups in variable are allowed. Must be a factor variable with exactly two levels.

Approximate Variance: Calculations to approximate the variance and confidence intervals are available.

Allowed Time Values: Allows only integer time.

Bounded Estimates: Estimates are guaranteed to be bounded in the 0 to 1 probability range.

Monotone Function: Estimates are not guaranteed to be monotone.

Dependencies: This method relies on the survtmle and SuperLearner packages.

TMLE is a two-step procedure. First, initial estimates for the treatment-assignment and the outcome-mechanisms are made using loss-based learning. This is implemented here using the SuperLearner methodology. In the next step, the estimates obtained by using the outcome-mechanism model are fluctuated based on information from the treatment-assignment model. If the outcome model is already consistent, this fluctuation is very small and the estimates stay consistent. If the outcome model is biased, the fluctuation removes the bias whenever the treatment assignment model is consistent. This process is iterative and continues until a threshold is hit (either the maximum number of iterations is reached or the bias is smaller than the specified tolerance, see ?survtmle).

As has been shown in multiple studies by Mark J. van der Laan and colleagues, this method has some desirable mathematical properties and generally performs well in appropriate scenarios. The biggest problem is however, that it is only defined for discrete (integer-valued) survival times. Simply discretizing continuous survival times only works to a certain extent and is generally discouraged.

When the sample size is large or many time points are of interest, this method will also be *very* slow. While possible to run, bootstrapping would take an enormous amount of time and is therefore discouraged.

Value

Adds the following additional objects to the output of the adjustedsurv function:

survtmle_object: The object created using the survtmle function.

survtmle.timepoints_object: The object created using the survtmle.timepoints function.

References

Ori M. Stitelman and Mark J. van der Laan (2010). "Collaborative Targeted Maximum Likelihood for Time to Event Data". In: The International Journal of Biostatistics 6.1

David Benkeser, Marco Carone, and Peter B. Gilbert (2018). "Improved Estimation of the Cumulative Incidence of Rare Outcomes". In: Statistics in Medicine 37.2, pp. 280-293

Megan S. Schuler and Sherri Rose (2017). "Targeted Maximum Likelihood Estimation for Causal Inference in Observational Studies". In: American Journal of Epidemiology 186.1, pp. 65-73

Author

The wrapper function was written by Robin Denz, the survtmle package (which this wrapper is based on) was written by David Benkeser and Nima Hejazi. See ?survtmle for more details.

See also

Examples

# not run because any meaningful example is too slow
if (FALSE) {
library(adjustedCurves)
library(survtmle)

set.seed(42)

# simulate some data as example
sim_dat <- sim_confounded_surv(n=50, max_t=1.2)
sim_dat$group <- as.factor(sim_dat$group)

# only works with integer time, only unbiased with small amounts of them
sim_dat$time <- round(sim_dat$time*15) + 1

# calculate adjusted survival curves, using SuperLearner but only
# using the SL.glm library. In practice you would want to use more than
# that. See ?survtmle
adjsurv <- adjustedsurv(data=sim_dat,
                        variable="group",
                        ev_time="time",
                        event="event",
                        method="tmle",
                        adjust_vars=c("x1", "x2", "x3", "x4", "x5", "x6"),
                        SL.ftime=c("SL.glm"),
                        SL.ctim=c("SL.glm"),
                        SL.trt=c("SL.glm"))

# plot the curves
plot(adjsurv)
}